2-2 Clear(Challenge Lv6)

+++ src/main/java/org/example/expert/domain/comment/service/CommentAdminService.java
@@ -2,6 +2,8 @@ package org.example.expert.domain.comment.service;
 
 import lombok.RequiredArgsConstructor;
 import org.example.expert.domain.comment.repository.CommentRepository;
+import org.example.expert.domain.common.dto.AuthUser;
+import org.example.expert.domain.user.enums.UserRole;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,7 +14,10 @@ public class CommentAdminService {
     private final CommentRepository commentRepository;
 
     @Transactional
-    public void deleteComment(long commentId) {
+    public void deleteComment(AuthUser authUser, long commentId) {
+        if (!authUser.getUserRole().equals(UserRole.ADMIN)) {
+            throw new IllegalArgumentException("관리자만 댓글을 삭제할 수 있습니다.");
+        }
         commentRepository.deleteById(commentId);
     }
 }
